First complete post-docker version
