--- sandbox/remote_test.php
@@ -1,21 +1,0 @@
-<form method="POST" action="remote.php" target="remote_result">
-<input type="text" name="password">
-<hr/>
-Code: <br/>
-<textarea style="width: 100%;" rows="5" name="code">
-#include &lt;stdio.h&gt;
-main() {
-  printf("Hello World\n");
-}
-</textarea>
-<hr/>
-Input: <br/>
-<textarea style="width: 100%;" name="input">
-</textarea>
-<hr/>
-Main: <br/>
-<textarea style="width: 100%;" name="main">
-</textarea>
-<hr/>
-<input type="submit">
-</form>
